refactor(stage-ui): use kebab-case file names
