+++ AGENTS.md
@@ -150,7 +150,7 @@ Concise but detailed reference for contributors working across the `moeru-ai/air
 
 ## Readability, Naming, and Comments
 
-- File names: camelCase.
+- Use kebab-case for all file names.
 - Prefer names that rely on the module boundary for context instead of repeating package, product, protocol, or transport prefixes inside every symbol. A well-named module should let exported functions use short action-first names; repeat the larger context only when the symbol crosses a boundary where that context is no longer obvious.
 - Name functions after the domain operation they perform, not after the implementation layer that happens to contain them. This keeps call sites readable after refactors and avoids names becoming stale when code moves between files.
 - Avoid names that encode multiple layers of ownership into one symbol. If a name needs several qualifiers to be understandable, reconsider the module boundary or introduce a clearer local concept.
