+++ apps/stage-web/src/main.ts
@@ -7,7 +7,7 @@ import NProgress from 'nprogress'
 import { autoAnimatePlugin } from '@formkit/auto-animate/vue'
 import { PiniaColada } from '@pinia/colada'
 import { isEnvTruthy } from '@proj-airi/stage-shared'
-import { trackButtonPlugin } from '@proj-airi/stage-ui/directives/trackButton'
+import { trackButtonPlugin } from '@proj-airi/stage-ui/directives/track-button'
 import { MotionPlugin } from '@vueuse/motion'
 import { createPinia } from 'pinia'
 import { setupLayouts } from 'virtual:generated-layouts'
